Extract coverbs, locatives, and posture predicates
